requirements.txt
RPi.GPIO==0.7.0
SpeechRecognition==3.8.1
pyttsx3==2.71
google-api-python-client==1.7.10
pyaudio==0.2.11
pyttsx3==2.71
python-i18n==0.3.6
